refactor(tools): migrate all tool logic to MCP servers, delete tools/ directory
**Key Architectural Changes**

1. **reflection_tool moved into BaseReActAgent**
   - Absorb reflection_tool from deleted tools/base_agent_tools.py
   - reflection_tool remains in agents/base_agent_template.py (needs InjectedState)
   - Automatically prepended to all agent tool lists

2. **Legislation Finder Agent - Inline Tool Adapters**
   - web_search: thin adapter calling search_legislation MCP + extract_search_results
   - reliability_analysis: thin adapter calling analyze_reliability MCP server
   - Both return LangGraph Command objects for state updates
   - Moved from tools/legislation_finder.py → agents/legislation_finder.py

3. **Political Commentary Agent - Inline Tool Adapters**
   - political_figure_finder: adapter calling find_political_figures MCP
   - search_political_commentary: adapter calling search_political_content + extract_commentary MCP
   - search_political_social_media: adapter calling search_politician_tweets MCP
   - All return LangGraph Command objects for state updates
   - Moved from tools/political_commentary_finder.py → agents/political_commentary_finder.py

4. **Deleted tools/ Directory**
   - tools/__init__.py (empty)
   - tools/base_agent_tools.py (reflection_tool moved to agents/)
   - tools/legislation_finder.py (adapters moved to agents/)
   - tools/political_commentary_finder.py (adapters moved to agents/)

5. **Updated Eval Test Imports**
   - evals/test_legislation_finder.py: tools.* → agents.*
   - evals/test_political_commentary.py: tools.* → agents.*
   - evals/test_e2e_pipeline.py: @patch paths updated
   - evals/conftest.py: removed Brave fixtures, cleaned up imports
   - evals/README.md: removed Brave references

Result: tools/ directory is completely eliminated. All business logic lives in MCP servers.
Agent files define inline adapters that call MCP clients and wrap results in Commands.
